update earn + dao services
